Seperate test container configurations due to abstraction. Integration tests starter.
