+++ src/test/java/com/ntloc/demo/customer/CustomerIntegrationTest.java
@@ -1,0 +1,30 @@
+package com.ntloc.demo.customer;
+
+import com.ntloc.demo.AbstractTestContainersTest;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class CustomerIntegrationTest extends AbstractTestContainersTest {
+
+    static final String CUSTOMER_API_PATH = "/api/v1/customers";
+
+    @Test
+    void createCustomer() {
+    }
+
+    @Test
+    void updateCustomer() {
+    }
+
+    @Test
+    void deleteCustomer() {
+    }
+}
