+++ src/test/java/com/ntloc/demo/customer/CustomerRepositoryTest.java
@@ -1,16 +1,12 @@
 package com.ntloc.demo.customer;
 
+import com.ntloc.demo.AbstractTestContainersTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 import java.util.Optional;
 
@@ -18,13 +14,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@Testcontainers
-class CustomerRepositoryTest {
-
-    @Container
-    @ServiceConnection
-    static PostgreSQLContainer<?> postgreSQLContainer =
-            new PostgreSQLContainer<>(DockerImageName.parse("postgres:16.2"));
+class CustomerRepositoryTest extends AbstractTestContainersTest {
 
     @Autowired
     private CustomerRepository customerRepository;
@@ -44,12 +34,6 @@ class CustomerRepositoryTest {
     @AfterEach
     void tearDown() {
         customerRepository.deleteAll();
-    }
-
-    @Test
-    void canEstablishDatabaseContainerConnection() {
-        assertTrue(postgreSQLContainer.isCreated(), "Database container should be created");
-        assertTrue(postgreSQLContainer.isRunning(), "Database container should be running");
     }
 
     @Test
